Move point addition and deletion to separate impl blocks
