+++ Rust/src/tree/mod.rs
@@ -9,5 +9,11 @@ pub use cut::Cut;
 mod node;
 pub use node::{Internal, Leaf, Node};
 
+mod tree_point_addition;
+pub use tree_point_addition::AddResult;
+
+mod tree_point_deletion;
+pub use tree_point_deletion::DeleteResult;
+
 mod tree;
-pub use tree::{AddResult, DeleteResult, NodeTraverser, Tree};
+pub use tree::{NodeTraverser, Tree};
